requirements.txt
groq
google-genai
python-dotenv
requests